refactor: move task cleanup to task details layout
